Optimize unit tests
- Split vm_dummy tests. Moved tests that only work in the Linux/amd64 environment to a separate file
- Adjust log levels or method to remove too verbose messages
- Modified not to show logs of successful tests.
