+++ contract/system/vprt_test.go
@@ -1,8 +1,10 @@
 package system
 
 import (
+	"bytes"
 	"container/list"
 	"fmt"
+	"github.com/aergoio/aergo-lib/log"
 	"math/big"
 	"math/rand"
 	"os"
@@ -23,6 +25,8 @@ const (
 )
 
 var (
+	logger = log.NewLogger("system.test")
+
 	valHundred = new(big.Int).SetUint64(100)
 	valTen     = new(big.Int).SetUint64(10)
 
@@ -170,7 +174,7 @@ func initRankTableRand(rankMax uint32) {
 func openSystemAccount(t *testing.T) *state.ContractState {
 	s, err := vprStateDB.OpenContractStateAccount(types.ToAccountID([]byte(types.AergoSystem)))
 	assert.NoError(t, err, "fail to open the system contract state")
-	fmt.Printf(
+	logger.Debug().Msgf(
 		"(after) state, contract: %s, %s\n",
 		enc.ToString(vprStateDB.GetRoot()),
 		enc.ToString(s.GetStorageRoot()))
@@ -395,8 +399,12 @@ func TestVprTotalPower(t *testing.T) {
 	assert.NoError(t, err, "fail to update the voting power ranking")
 
 	votingPowerRank.checkValidity(t)
-	votingPowerRank.voters.dump(os.Stdout, 10)
-	votingPowerRank.voters.dump(os.Stdout, 0)
+	if logger.IsDebugEnabled() {
+		buffer := &bytes.Buffer{}
+		votingPowerRank.voters.dump(buffer, 10)
+		votingPowerRank.voters.dump(buffer, 0)
+		t.Log(buffer)
+	}
 }
 
 func TestVprSingleWinner(t *testing.T) {
@@ -416,7 +424,7 @@ func TestVprSingleWinner(t *testing.T) {
 	}
 
 	for addr, count := range stat {
-		fmt.Printf("%v: pwr = %v, wins # = %v\n",
+		logger.Debug().Msgf("%v: pwr = %v, wins # = %v\n",
 			addr, votingPowerRank.votingPowerOf(addr), count)
 	}
 }
@@ -438,7 +446,7 @@ func TestVprPickWinner(t *testing.T) {
 	}
 
 	for addr, count := range stat {
-		fmt.Printf("%v: pwr = %v, wins # = %v\n",
+		logger.Debug().Msgf("%v: pwr = %v, wins # = %v\n",
 			addr, votingPowerRank.votingPowerOf(addr), count)
 	}
 }
